L1 work on bugs
