+++ lesson1/fruit/Apple.java
@@ -1,0 +1,7 @@
+package lesson1.fruit;
+
+public class Apple extends Fruit {
+    public Apple() {
+        super.weight = 1.0f;
+    }
+}
