+++ lesson1/fruit/Fruit.java
@@ -1,0 +1,12 @@
+package lesson1.fruit;
+
+/*Есть классы Fruit -> Apple, Orange;(больше фруктов не надо)*/
+public abstract class Fruit {
+    float weight;
+
+    public float getWeight() {
+        return weight;
+    }
+}
+
+
