+++ lesson1/fruit/Orange.java
@@ -1,0 +1,7 @@
+package lesson1.fruit;
+
+public class Orange extends Fruit {
+    public Orange() {
+        super.weight = 1.5f;
+    }
+}
